Refactor verify flow
