+++ internal/verify/__init__.py
@@ -1,0 +1,9 @@
+from .verifier import TMTVerifyIssue, TMTVerifyIssueType, Verifier
+from .verdicts import VerdictsVerifier
+
+__all__ = [
+    "TMTVerifyIssue",
+    "TMTVerifyIssueType",
+    "Verifier",
+    "VerdictsVerifier"
+]
